refactor: ts of expity text
