+++ packages/trader/src/Modules/Trading/Components/Form/TradeParams/Duration/__tests__/expiry-text.spec.tsx
