+++ packages/trader/src/Modules/Trading/Components/Form/TradeParams/Duration/index.ts
